Support redirects from Professorships site
The legacy D7 site: professorships.stanford.edu is being redirected to giving.stanford.edu/endowed-positions which is working well, but redirects to the "sub-pages" are breaking due to the D7 site using spaces in the urls. See http://professorships.stanford.edu/content/Wu%20Tsai%20Neurosciences%20Institute as an example. The proposed dynamic redirect is intended to catch all traffic redirected from the legacy urls and direct users to the top-level landing page for the Endowed Positions content at /endowed-positions.